First Component with @Component
A component contains a template, data and logic that work together
to form a part of a DOM tree.

In this exercise, we reconfigure the app.component.ts file from scratch
in order to review the imports, decorator and class that comprises a
component class.

An Angular component is a TypeScript file that contains the following:

* An import of Component from @angular/core
* A @Component decorator that decorates the class with the following
  properties:
  - selector (This selector is insert in the index.html file as a tag
    to denote where the Angular application component should be
    rendered in the page). <app-root> is used by default.
  - styleUrls (An array of stylesheets used to create the presentation
    layer of the component
  - template or templateUrl (Used to mark up the HTML for the component
    inline or in an external file)
* An exported class named AppComponent